fix(foundry_script): Apply int's 32-bit width uniformly, removing the carve-out
`_applied_numeric_type()` no longer maps `int`'s INT32 registry descriptor to
NONE, so the scripting-language `int` spelling behaves exactly like `uint`,
`long`, and `ulong`: it narrows at assignment, participates in the carrier-
mixing diagnostic instead of the generic operand message, and constrains
typed containers and generic type arguments to their exact 32-bit range.

Two analyzer-adjacent bugs surfaced while widening `int` uniformly and are
fixed alongside the carve-out removal, since they blocked a clean run:

- `int(value)` constructor calls lost their width because the call-typing
  path used the carrier-only `get_builtin_type()` lookup and, for the
  non-constant-argument branch, re-derived the result type from the engine's
  own constructor list, which carries no FoundryScript width metadata.
- Compiling a `const X: Variant = Crate[int]` class-handle constant resolved
  its `int` type argument through a separate, ad hoc identifier lookup that
  never consulted the width registry, so the constant lost the width its
  direct-annotation counterpart already had.

The resulting fixture and test churn is extensive but mechanical: every
`-> int` function whose body genuinely returns a native 64-bit value (mostly
`.size()`/`.length()`) now declares `-> long`, matching what it already
returned; a handful of C++ tests that hand-construct `ContainerType`/
`FSDataType` values for `int` needed their `numeric_type` filled in to match
the registry; and one cross-script trait-witness scenario is pinned to its
new, honest failure because trait witnesses are matched through the compiled
`PropertyInfo`/`MethodInfo` boundary, which still transports a scalar
integer's carrier only, not its declared width — recovering that is tracked
as a follow-up rather than folded into this change.

Closes #1684

+++ modules/foundry_script/tests/scripts/analyzer/errors/named_call_argument_rest_parameter.fs
@@ -1,4 +1,4 @@
-func collect(first: int, ...rest: Array) -> int:
+func collect(first: int, ...rest: Array) -> long:
 	return first + rest.size()
 
 func test():
