+++ modules/foundry_script/tests/scripts/runtime/features/typed_rest_parameter_concrete.fs
@@ -1,25 +1,25 @@
 class Marker:
 	var id := 0
 
-func collect(...values: Array[int]) -> int:
+func collect(...values: Array[int]) -> long:
 	Utils.check(values.is_typed())
 	Utils.check(values.get_typed_builtin() == TYPE_INT)
 	return values.size()
 
-func collect_floats(...values: Array[float]) -> int:
+func collect_floats(...values: Array[float]) -> long:
 	# Surplus arguments follow the fixed-parameter conversion policy, so passing ints stores floats.
 	for value in values:
 		Utils.check(typeof(value) == TYPE_FLOAT)
 	return values.size()
 
-func collect_nullable(...values: Array[Marker?]) -> int:
+func collect_nullable(...values: Array[Marker?]) -> long:
 	# A nullable element cannot be expressed as a typed container, so the collected array is erased
 	# exactly like an ordinary `Array[Marker?]` local.
 	var ordinary: Array[Marker?] = []
 	Utils.check(values.is_typed() == ordinary.is_typed())
 	return values.size()
 
-func collect_gradual(...values: Array) -> int:
+func collect_gradual(...values: Array) -> long:
 	Utils.check(!values.is_typed())
 	return values.size()
